abstract eots mgr func
